refactor: move check_legacy_config and ensure_dirs to utils module
